requirements.txt
Django==3.0.6
python-decouple>=3.1
dj-database-url>=0.5.0
dj-static>=0.0.6
psycopg2==2.7.7

appdirs==1.4.3
asgiref==3.2.7
astroid==2.3.3
beautifulsoup4==4.9.0
colorama==0.4.3
distlib==0.3.0
dj-config-url==0.1.1
django-admin-sortable2==0.7.6
django-crispy-forms==1.9.0
django-dynamic-formsets==0.0.8
django-heroku==0.3.1
django-smart-selects==1.5.4
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.4
pytz==2019.3
selenium==3.141.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.16
whitenoise==5.0.1
wrapt==1.11.2
xlwt==1.3.0
zipp==3.1.0
